requirements.txt
langchain-google-genai~=2.1.2
langgraph~=0.3.21
typing_extensions~=4.13.0
pillow~=11.1.0
requests~=2.32.3
soundfile~=0.13.1
ipython~=9.0.2
kokoro~=0.9.2
langchain-core~=0.3.49